# perf: cache intl object